face-service/requirements.txt
# Face Verification Service - Requirements
# Install: pip install -r requirements.txt
# Compatible with Python 3.14+

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Face Recognition (InsightFace)
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.2.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.7

# Rate Limiting & Utils
slowapi>=0.1.9
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
